configurable `tts_sample_rate` + clean up
